522: v1.4: add tasks and modified pre_luncher
